[clang] Refactor of creating method signatures. First step to use the map of pointers to find method declarations.
Summary:
Refactor of creating method signatures. First step to use the map of pointers to find method declarations.
The idea is to have a function that creates a method signature from a declaration, later we can get the declaration from a pointer
and use this function to retrieve the method signature.

(authored by @dulmarod)