Remove variable resolution and use pointers to declarations instead
Summary: @​public
This removes the old way of finding variable declarations to create sil variables and replaces it with
a a new way based on the map from pointers to declarations.
Basically, every variable dereference contains a pointer to the variable declaration, with that we can
build the corresponding sil variable.

Reviewed By: @akotulski

Differential Revision: D2536000

fb-gh-sync-id: dd29cf9