Fixing mangling of captured variables
Summary: public This diff fixes incorrect mangling of captured variables in blocks. Because they are formals,
they shouldn't be mangled, but this case was not taken into account. This caused an assert false in the
example infer/tests/codetoanalyze/objc/frontend/block/block.m which wasn't caught before because there
wasn't an endtoend test for it.

Reviewed By: akotulski

Differential Revision: D2560379

fb-gh-sync-id: db500b6