Adding static variables to blocks class to get retain cycles
Summary: public
The static and global variables used in blocks don't appear in the ast as captured.
We need them however to try and find retain cycles involving those variables.
This diff adds a way of collecting the static variables used in blocks and treat them
like we treat other captured variables to find retain  cycles.

Reviewed By: ddino

Differential Revision: D2663727

fb-gh-sync-id: d5b44ec