Don't crash when method decl is not found
Summary: public
Not finding C++ method in decl map was considered impossible to happen.
It turns out it can happen because ast exporter wasn't exporting template
instantiations.

So instead, pick a default when method decl is not found. This will be right in
most cases and will allow us to analyze rest of the function without crashing
frontend

Reviewed By: dulmarod

Differential Revision: D2695653

fb-gh-sync-id: 4077c59