Add support for passing blacklist regex for files to not analyze with Buck
Summary: public
With this change it will be possible to instruct Buck to not analyze certain files whose names match
with the regex passed in input.
So for example if you want to skip the analysis of files containing `skip_me` on their names, then
the command:
  infer --use-flavors --blacklist-regex ".*skip_me.*" -- buck build <target-name>
will do it, because the `.*skip_me.*` regex will match the desired rule.

Reviewed By: jvillard

Differential Revision: D2702872

fb-gh-sync-id: 498b17b