Skip the Buck build command when the list of target is empty
Summary: public
Since we are only running the analysis on the Buck targets that are of type java_library or android_library, it may happen that the filtering gives an empty list of targets.

This diff does two things:
 1) it filters out the empty strings `''` that can be returned when parsing the `buck query` command, and
 2) skip the analysis when the list of targets to analyze is empty.

It also fixes the debug output for the list of targets to analyze, which was previously always printing the list of targets passed in the command line argument, which led for this issue to be unoticed even though it may have been there for a while already.

Reviewed By: sblackshear

Differential Revision: D2704949

fb-gh-sync-id: 83c62d8