Adding builtins for getters and calling them by skip
Summary: public
We were generating getters and setters in the frontend, and then removing them if they were not needed
in the preanalysis. This diff adds a builtin getter that gets called if we are going to skip the function. That
means, if there was a getter written by the developers we still use that one rather than the builtin.
Code for setter and cleanup of preanalysis will come in a next diff.

Reviewed By: jvillard

Differential Revision: D2702890

fb-gh-sync-id: d65f25e