treat guava preconditions checks as assume instead of exeption throwing assertions
Summary:
public
The case when a resource leaks is reported because the the resource was not closed on the execution branch created by the preconditions checks are not very interesting in practice because the exceptions thrown, either `NullPointerException` or `IllegalStateException` are very rarely caught anyway. So the legimate use of preconditions checks is creating spurious resource leak reports.

Reviewed By: sblackshear

Differential Revision: D2707227

fb-gh-sync-id: 6aece73