Add correct ivar name to the getters and save the ivar to the tenv if not available
Summary:
public The ivar corresponding to the property is only available in the ast when the
implementation of the peroperty is available. Otherwise we add an ivar with the correct type
and the default name to the tenv and use it in the getter (and later in the setter).
This was not causing crashes because the generated code was swallowing the Missing_fld exception.
Now it flags it.

Reviewed By: akotulski

Differential Revision: D2734217

fb-gh-sync-id: 21c62af