better instructions for first-time users
Summary:
public
The instructions to install opam from the web on linux were not enough:
you need to install opam somewhere on the PATH. Moreover, the wget
command could erase "opam" if launched from the infer repo, which
already has an "opam" file.

./build-infer.sh's "opam init" command would fail if ocaml was not already
installed, instead of installing a default compiler.

Finally, check that ocamlfind is installed using the same logic as other
ocamlfind packages, to get an error message explaining how to get it
using opam if it's not installed, instead of just a failure.

Reviewed By: jeremydubreil

Differential Revision: D2760678

fb-gh-sync-id: 678178e