distinguishing between failures and timeouts
Summary: public Crashes during the analysis are classified as timeouts in the .specs file. In addition, when there is a timeout, it does not say *why* the timeout occurred (hard time, symops, or recursion). This diff adds this information to the .specs file and adds a "fail hard" mode where crashes and timeouts will actually stop the analysis in developer mode (but will still be hidden in the normal production mode).

Reviewed By: jeremydubreil

Differential Revision: D2725382

fb-gh-sync-id: b0b4e5e