Use the Mangled module to name the parameters in the procudure description instead of simple string
Summary:
public
The paramtere where defined as simple strings in the procedure description. This diff force the use of the Mangled module to avoid possible conflict when converting variable back and forth from string to pvar. The code is now more consistent as the local variable were already named using mangled names.

Reviewed By: jberdine

Differential Revision: D2782863

fb-gh-sync-id: 1867574