Change the way nested structs are translated
Summary:
public
Reuse logic in cFrontend.ml to deal with nested structs. This allows to have less duplicated logic which will result
in less bugs.
As a bonus, it fixes the problem of nested template class (which wasn't be handled right).
And one random fix in c_type -> sil_type conversion (I'm amazed it worked before)

Reviewed By: dulmarod

Differential Revision: D2773687

fb-gh-sync-id: e312599