Fix the buck integration for the analysis of Buck itself
Summary:
public
The analysis of the Buck project was failing because this script converts the aliases into their expanded target names. It turns out that for Buck, the name of the command is `buck` and the name of the alias is also `buck`, which led to conflicts. This is now fixed by only replacing the targets in the part of the command line that is after `build` in `buck build ...`

Reviewed By: sblackshear

Differential Revision: D2914446

fb-gh-sync-id: ddce4ed
shipit-source-id: ddce4ed