Add support for @NoAllocation in the performance critical checker
Summary:public
Add to the code to detect violation of the `NoAllocation` annotation. This diff adds the code to detect such issue based on the code of the `PerformanceCritical` checker. In the next diff, I will refine the list of acceptable allocations, like new exceptions, etc, and add the list of corresponding tests.

Reviewed By: sblackshear

Differential Revision: D2938641

fb-gh-sync-id: 9a047dd
shipit-source-id: 9a047dd