Warn on unused identifiers
Summary:public
I have seen enough comments in this space by people during code review to switch on the analyses the compiler can already do.  This diff is an automated renaming of unused identifiers to _, with a few additional changes made when reading the diff of the results for things that stood out as particularly strange.  This base-lines all of the existing warnings.  I'm not sure this is a good idea, since it might be better for those familiar with each part of the code to look at these warnings and use them as pointers to suspicious code.

Reviewed By: jeremydubreil

Differential Revision: D2938376

fb-gh-sync-id: 6e67817
shipit-source-id: 6e67817