Add byte target
Summary:public
This is intended as a target to build during development when quick
compilation and no duplicate compilation warnings are desired.

This builds byte code for InferAnalyze, InferPrint, InferJava, and
InferClang.  This is one ocamlbuild invocation, so builds (and reports
warnings on) each source file once.  Building bytecode is currently
about twice as fast as native code.

The llvm frontend is currently not built by this since it requires
menhir and standard ocamlyacc is required by the backend, and ocamlbuild
can only use one at a time.

Reviewed By: jvillard

Differential Revision: D2953050

fb-gh-sync-id: 2e90fcd
shipit-source-id: 2e90fcd